+++ XRViewer/Resources/XRViewer-Bridging-Header.h
@@ -9,3 +9,10 @@
 #import "Prefix.h"
 #import "ARKController.h"
 #import "Animator.h"
+
+typedef NS_ENUM(NSInteger, ResetTrackingOption) {
+    ResetTracking,
+    RemoveExistingAnchors,
+    SaveWorldMap,
+    LoadSavedWorldMap
+};